refactor: Refactor js file
